backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
yfinance==0.1.63
groq==0.4.0
python-dotenv==0.19.0